fix: support float fps for accurate video frame rate handling
- Change fps parameter type from int to float in audio_processor, inference, and realtime_inference
- Auto-detect video FPS from source file in realtime_inference mode
- Fixes frame sync issues with non-integer frame rates (e.g., 29.97, 23.976)

+++ musetalk/utils/audio_processor.py
@@ -58,7 +58,7 @@ class AudioProcessor:
         # Trim the last segment to remove padding
         sr = 16000
         audio_fps = 50
-        fps = int(fps)
+        fps = float(fps)
         whisper_idx_multiplier = audio_fps / fps
         num_frames = math.floor((librosa_length / sr) * fps)
         actual_length = math.floor((librosa_length / sr) * audio_fps)
